+++ src/main/java/store_order_resource_outwards_mgmt/controller/read/StoreOrderResourceOutwardsReadPublic_Controller.java
@@ -60,7 +60,7 @@ public class StoreOrderResourceOutwardsReadPublic_Controller
 
 
 	@GetMapping(value = "/getAllRowsForMode/{md}", produces = { MediaType.APPLICATION_JSON_VALUE })
-	public ResponseEntity<CopyOnWriteArrayList<StoreOrderResourceOutward_DTO>> getAllRowsForMode(@PathVariable Short md) 
+	public ResponseEntity<CopyOnWriteArrayList<StoreOrderResourceOutward_DTO>> getAllRowsForMode(@PathVariable Integer md) 
 	{
 		CompletableFuture<CopyOnWriteArrayList<StoreOrderResourceOutward_DTO>> stMaster_DTOs = null;
 		CopyOnWriteArrayList<StoreOrderResourceOutward_DTO> stMasterList = null;
@@ -150,7 +150,7 @@ public class StoreOrderResourceOutwardsReadPublic_Controller
 	}
 	
 	@GetMapping(value = "/getTotalRequestedForJob/{jid}/{rid}/{mid}", produces = { MediaType.APPLICATION_JSON_VALUE })
-	public ResponseEntity<Float> getTotalRequestedForJob(@PathVariable Long jid, @PathVariable Long rid, @PathVariable Short mid)
+	public ResponseEntity<Float> getTotalRequestedForJob(@PathVariable Long jid, @PathVariable Long rid, @PathVariable Integer mid)
 	{
 		CompletableFuture<Float> qty = null;
 		Float result = null;
@@ -168,7 +168,7 @@ public class StoreOrderResourceOutwardsReadPublic_Controller
 	}
 	
 	@GetMapping(value = "/getTotalQtyRequestedBeforeThisRequest/{sid}/{rid}/{mid}", produces = { MediaType.APPLICATION_JSON_VALUE })
-	public ResponseEntity<Float> getTotalQtyRequestedBeforeThisRequest(@PathVariable Long sid, @PathVariable Long rid, @PathVariable Short mid)
+	public ResponseEntity<Float> getTotalQtyRequestedBeforeThisRequest(@PathVariable Long sid, @PathVariable Long rid, @PathVariable Integer mid)
 	{
 		CompletableFuture<Float> qty = null;
 		Float result = null;
@@ -186,7 +186,7 @@ public class StoreOrderResourceOutwardsReadPublic_Controller
 	}
 	
 	@GetMapping(value = "/getTotalQtyRequestedBeforeDTTM/{dt}/{rid}/{mid}", produces = { MediaType.APPLICATION_JSON_VALUE })
-	public ResponseEntity<Float> getTotalQtyRequestedBeforeDTTM(@PathVariable String dt, @PathVariable Long rid, @PathVariable Short mid)
+	public ResponseEntity<Float> getTotalQtyRequestedBeforeDTTM(@PathVariable String dt, @PathVariable Long rid, @PathVariable Integer mid)
 	{
 		CompletableFuture<Float> qty = null;
 		Float result = null;
@@ -222,7 +222,7 @@ public class StoreOrderResourceOutwardsReadPublic_Controller
 	}
 	
 	@GetMapping(value = "/getTotalAllocatedForJob/{jid}/{rid}/{mid}", produces = { MediaType.APPLICATION_JSON_VALUE })
-	public ResponseEntity<Float> getTotalAllocatedForJob(@PathVariable Long jid, @PathVariable Long rid, @PathVariable Short mid)
+	public ResponseEntity<Float> getTotalAllocatedForJob(@PathVariable Long jid, @PathVariable Long rid, @PathVariable Integer mid)
 	{
 		CompletableFuture<Float> qty = null;
 		Float result = null;
@@ -240,7 +240,7 @@ public class StoreOrderResourceOutwardsReadPublic_Controller
 	}
 	
 	@GetMapping(value = "/getTotalQtyAllocatedBeforeThisRequest/{sid}/{rid}/{mid}", produces = { MediaType.APPLICATION_JSON_VALUE })
-	public ResponseEntity<Float> getTotalQtyAllocatedBeforeThisRequest(@PathVariable Long sid, @PathVariable Long rid, @PathVariable Short mid)
+	public ResponseEntity<Float> getTotalQtyAllocatedBeforeThisRequest(@PathVariable Long sid, @PathVariable Long rid, @PathVariable Integer mid)
 	{
 		CompletableFuture<Float> qty = null;
 		Float result = null;
@@ -258,7 +258,7 @@ public class StoreOrderResourceOutwardsReadPublic_Controller
 	}
 	
 	@GetMapping(value = "/getTotalQtyAllocatedBeforeDTTM/{dt}/{rid}/{mid}", produces = { MediaType.APPLICATION_JSON_VALUE })
-	public ResponseEntity<Float> getTotalQtyAllocatedBeforeDTTM(@PathVariable String dt, @PathVariable Long rid, @PathVariable Short mid)
+	public ResponseEntity<Float> getTotalQtyAllocatedBeforeDTTM(@PathVariable String dt, @PathVariable Long rid, @PathVariable Integer mid)
 	{
 		CompletableFuture<Float> qty = null;
 		Float result = null;
@@ -294,7 +294,7 @@ public class StoreOrderResourceOutwardsReadPublic_Controller
 	}
 	
 	@GetMapping(value = "/getTotalBookedForJob/{jid}/{rid}/{mid}", produces = { MediaType.APPLICATION_JSON_VALUE })
-	public ResponseEntity<Float> getTotalBookedForJob(@PathVariable Long jid, @PathVariable Long rid, @PathVariable Short mid)
+	public ResponseEntity<Float> getTotalBookedForJob(@PathVariable Long jid, @PathVariable Long rid, @PathVariable Integer mid)
 	{
 		CompletableFuture<Float> qty = null;
 		Float result = null;
@@ -312,7 +312,7 @@ public class StoreOrderResourceOutwardsReadPublic_Controller
 	}
 	
 	@GetMapping(value = "/getTotalQtyBookedBeforeThisRequest/{sid}/{rid}/{mid}", produces = { MediaType.APPLICATION_JSON_VALUE })
-	public ResponseEntity<Float> getTotalQtyBookedBeforeThisRequest(@PathVariable Long sid, @PathVariable Long rid, @PathVariable Short mid)
+	public ResponseEntity<Float> getTotalQtyBookedBeforeThisRequest(@PathVariable Long sid, @PathVariable Long rid, @PathVariable Integer mid)
 	{
 		CompletableFuture<Float> qty = null;
 		Float result = null;
@@ -330,7 +330,7 @@ public class StoreOrderResourceOutwardsReadPublic_Controller
 	}
 	
 	@GetMapping(value = "/getTotalQtyBookedBeforeDTTM/{dt}/{rid}/{mid}", produces = { MediaType.APPLICATION_JSON_VALUE })
-	public ResponseEntity<Float> getTotalQtyBookededBeforeDTTM(@PathVariable String dt, @PathVariable Long rid, @PathVariable Short mid)
+	public ResponseEntity<Float> getTotalQtyBookededBeforeDTTM(@PathVariable String dt, @PathVariable Long rid, @PathVariable Integer mid)
 	{
 		CompletableFuture<Float> qty = null;
 		Float result = null;
@@ -367,7 +367,7 @@ public class StoreOrderResourceOutwardsReadPublic_Controller
 	}
 	
 	@GetMapping(value = "/getTotalQtyMovedForResourceForJob/{jid}/{rid}/{mid}", produces = { MediaType.APPLICATION_JSON_VALUE })
-	public ResponseEntity<Float> getTotalQtyMovedForResourceForJob(@PathVariable Long jid, @PathVariable Long rid, @PathVariable Short mid)
+	public ResponseEntity<Float> getTotalQtyMovedForResourceForJob(@PathVariable Long jid, @PathVariable Long rid, @PathVariable Integer mid)
 	{
 		CompletableFuture<Float> qty = null;
 		Float result = null;
@@ -385,7 +385,7 @@ public class StoreOrderResourceOutwardsReadPublic_Controller
 	}
 	
 	@GetMapping(value = "/getTotalQtyMovedForResourceBeforeThisRequest/{sid}/{rid}/{mid}", produces = { MediaType.APPLICATION_JSON_VALUE })
-	public ResponseEntity<Float> getTotalQtyMovedForResourceBeforeThisRequest(@PathVariable Long sid, @PathVariable Long rid, @PathVariable Short mid)
+	public ResponseEntity<Float> getTotalQtyMovedForResourceBeforeThisRequest(@PathVariable Long sid, @PathVariable Long rid, @PathVariable Integer mid)
 	{
 		CompletableFuture<Float> qty = null;
 		Float result = null;
@@ -403,7 +403,7 @@ public class StoreOrderResourceOutwardsReadPublic_Controller
 	}
 	
 	@GetMapping(value = "/getTotalRowsMovedForResourcesBeforeThisDTTM/{dt}/{rid}/{mid}", produces = { MediaType.APPLICATION_JSON_VALUE })
-	public ResponseEntity<Float> getTotalRowsMovedForResourcesBeforeThisDTTM(@PathVariable Long sd, @PathVariable Long rid, @PathVariable Short mid)
+	public ResponseEntity<Float> getTotalRowsMovedForResourcesBeforeThisDTTM(@PathVariable Long sd, @PathVariable Long rid, @PathVariable Integer mid)
 	{
 		CompletableFuture<Float> qty = null;
 		Float result = null;
@@ -421,7 +421,7 @@ public class StoreOrderResourceOutwardsReadPublic_Controller
 	}
 
 	@GetMapping(value = "/getTotalRowsForResourcesBeforeThisDTTM/{dt}/{rid}/{mid}", produces = { MediaType.APPLICATION_JSON_VALUE })
-	public ResponseEntity<Float> getTotalRowsForResourcesBeforeThisDTTM(@PathVariable String dt, @PathVariable Long rid, @PathVariable Short mid)
+	public ResponseEntity<Float> getTotalRowsForResourcesBeforeThisDTTM(@PathVariable String dt, @PathVariable Long rid, @PathVariable Integer mid)
 	{
 		CompletableFuture<Float> qty = null;
 		Float result = null;
